Allow for setting of docker and docker-compose binaries.
When Fedora 32 was released, Docker was not compatible with it [1]. And
so, since then I've been using podman [2] as a drop-in-replacement.
Nowadays docker is fully compatible with Fedora 35 [3], as it now
supports CGroupsV2 and NFTables. However, this seems like a real bad
time to get to install it and move from a stable environment.

[1] - https://fedoramagazine.org/docker-and-fedora-32/
[2] - https://podman.io
[3] - https://fedoramagazine.org/docker-and-fedora-35/